feat(AAS-3): pr fixes and cleanup
