+++ frontend/src/components/images/ImageUploadButton.tsx
