+++ frontend/src/components/workspace/DeleteWorkspaceButton.tsx
@@ -3,19 +3,10 @@ import { useDeleteWorkspace } from '../../../generated'
 
 type DeleteWorkspaceButtonProps = {
   workspaceId: number
-  onDeleted?: () => void // optional callback after deletion
 }
 
-export function DeleteWorkspaceButton({ workspaceId, onDeleted }: DeleteWorkspaceButtonProps) {
-  const { mutate, error, isSuccess } = useDeleteWorkspace(
-    workspaceId,
-    {},
-    {
-      onSuccess: () => {
-        if (onDeleted) onDeleted()
-      },
-    },
-  )
+export function DeleteWorkspaceButton({ workspaceId }: DeleteWorkspaceButtonProps) {
+  const { mutate, error } = useDeleteWorkspace(workspaceId)
 
   return (
     <>
@@ -31,7 +22,6 @@ export function DeleteWorkspaceButton({ workspaceId, onDeleted }: DeleteWorkspac
       {error && (
         <p className="text-red-600 text-sm mt-1">{error.message || 'Failed to delete workspace'}</p>
       )}
-      {isSuccess && <p className="text-green-600 text-sm mt-1">Workspace deleted successfully!</p>}
     </>
   )
 }
