--- frontend/src/components/workspace/getWorkspaceUsers.tsx
@@ -1,13 +1,0 @@
-import { User } from '@/lib/types'
-
-export async function getWorkspaceUsers(workspaceId: number): Promise<User[]> {
-  //TODO: add in hooks when they come out
-
-  const workspaceIdStr = String(workspaceId)
-  const res = await fetch(`/api/workspaces/${workspaceIdStr}/users`)
-  if (!res.ok) {
-    throw new Error(`Failed to fetch users for workspace ${workspaceIdStr}`)
-  }
-  const data = (await res.json()) as User[]
-  return data
-}
